BugFix: various issues found by C.I. processes
Fixed a spurious open C type multi-line comment at start of:
3rdpart/communi/3rparty/mozilla/rdf_utils.c where I had chosen the GPL
only licence and rewritten the header in a manner consistant with the
Mozilla terms provided therein.

With the addition of Q_OBJECT macro into KeyUnit and Host classes
(to allow them to use Qt SIGNALs/SLOTs) it also seems to be necessary
to #include <Q_OBJECT> for some build platforms even though it must be
inherited from one of the other #include s...

Not being totally familiar with the methodology in adding overloaded
helpers for data items to QDataStream::operator<< and
QDataStream::operator>> friend type helpers I did not realise I had to
return the (first) argument reference for the two argument methods.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>