BugFix: module creation can now include sub-directories
Also includes error/success reporting (as the package exporter dialogue
no longer disappears as soon as "Export" is clicked).

Code refactored to support both 0.1x and 1.x versions of libzip - former
is currently only option on the current Linux Travis C.I. {uUbuntu Trusty
Tahr LTS} platform but later forms have error handling code that is
resistant to the string buffer overflow vulnerabilities in the older code.

Includes option to use a properly "temporary" staging area that will
dispose of itself once the module has been zipped into an archive file but,
to ensure compatibility with previous implementations it is not the default
option each time a module is made.

There is scope to enhance the use of the config.lua file to allow
additional details besides the module name to be stored in that file. I
have left TODO: markers where such detail could be processed.  In the past
I have experimented with code that included the following (optional data):
* Version (Symantic form?)
* Author name
* Email
* URL
* Creation Date (auto filled in)
* Last Saved Date (for synced modules)

however that is beyond the scope of this PR and requires additional
handling in the Host class at least.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>