Enhance: add Coverity Badge to collection...! [ci skip]
Since we have been added other badges, why not this one!

Note that since this is a non-code change I have set it not to invoke the
CI stages - however I think this means it will need an Admin to allow it to
be merged without CI pass reports!

Have also taken the liberty of emending the non-existent README entry in
the list in the qmake project file under "Other files" to README.md so that
the file edited herein can be seen and edited from within Qt Creator
easily.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>