Refactor: de-spaghetti-fy toolbar & editor treeWidget icon sizing code
The code that sized/resized the main (and replay) and editor toolbar icons
and editor treeWidget icons was inconsistent - when adjusted in the
preferences it only adjusted the editor of the current profile even if
others were also present.  Also some methods had an argument to receive
the size setting but ignored it and read the value directly from the
variable that stored it in the mudlet class.

Also:
* converted some tabs in the source code to spaces.
* added mudlet::setToolbarVisible(bool) and mudlet::setMenuBarVisible(bool)
  to simplify the coding when the relevant mudlet class items had their
  visibility adjusted in the preferences dialog.
* added code to ensure replay toolbar also gets it's icons adjusted when
  the main one to which it is docked gets adjusted.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>