Refactor: allow Preferences to be usable without a profile loaded
I anticipate a need to be able to show the options/preferences dialog on
start up when there is NO profile loaded.  This commit aims to make it
possible to do so.  It moves all the options that are not tied to a
specific Host instance to the first tab - and moves any that are to
elsewhere.  It then disables the other tabs so they cannot be accessed.

Internally it also refactors the diaProfilePreferences class so that ALL
methods check for and are safe to use even if dlgProfilePreferences::mpHost
is a nullptr.

To allow debugging controls to still be placed onto the dialog the
groupBox_Debug has been duplicated so that there is a groupBox_debugApplication
on the first tab and a groupBox_debugProfile on the last - at present they
are both hidden.

Accordingly the mudlet::show_options_dialog() slot no longer bails out if
there is not an "active" profile.

The indentation in a couple of places in dlgProfilePreferences.cpp are left
wrong so that the changes are more clearly visible - this will be updated
in the next commit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>