Fix mapper not receiving mouse events while sharing space with Labels
Added workaround for the mapper not receiving mouse events while
sharing space with a Label. This commit checks mouse click events in
Labels before allowing the Label to process the event. If the topmost
element at the cursor position is a child of the QWidget with
objectName "mapper", the event is manually redirected to that element.
This commit also adds overrides for wheelEvent and mouseMoveEvent to
 the TLabel class, for redirection purposes.