Revise: add mudlet class signals for Host creation and destruction
These will be needed to allow the Profile Preferences Dialogue to handle
the addition of a (first real) Host instance and (potentially, should other
developments happen) the removal of a (last real) Host instance.

Also fixed a corner case where a request to HostManager to delete a Host
when the specified named Host does not actually exist - it was still
returning a true even when that is not appropriate.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>