Remove version information entirely
AppImage validation is still complaining with:

mudlet.desktop: error: value "1.1" for key "Version" in group "Desktop Entry" is not a known version
ERROR: Desktop file contains errors. Please fix them. Please see
       https://standards.freedesktop.org/desktop-entry-spec/latest/
       for more information.

That is despite the standard clearly saying 1.1 is what should be used:

Version of the Desktop Entry Specification that the desktop entry conforms with. Entries that confirm with this version of the specification should use 1.1. Note that the version field is not required to be present.