Add QEvent data to TLabel callbacks
This commit reverts previous changes in TLabel.cpp that passed
additional arguments to the appropriate event handler function.
Instead, that functionality was moved to TLuaInterpreter. An
additional argument was declared for
TLuaInterpreter::callEventHandler, allowing for the passing of
a QEvent. This argument was given a default value to maintian
compatibility.

Now, in TLuaInterpreter::callEventHandler, the function
processes the passed in event for relevant information,
pushing it on to the Lua stack as a single Lua table. This
table is appended to the current list of parameters passed back
through the event function, which prevents breaking existing
Mudlet scripts.