Revise: following peer review (and bug detection)
The term used to identify the specifier for modules and packages in the
Lua commands to load and unload them have been revised to sound more
understandable to non-coders.

Whilst investigating the above it became clear that the "wrong" form of
error messages were being created for the "nil + error message"  type of
"run-time" errors; and an extra flag argument was added to
mudlet::unzip(...) to cause those (indirect, which was why it was not
immediately obvious before) cases.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>