Additional label callbacks (#1411)
* Add more mouse label callbacks

Added more mouse label callbacks, including wheel events,
double clicks, and move events. The wheel event appends two arguments
to the lua function call, corresponding to the x and y scroll
distance. The move event appends a variable number of arguments,
consisting of the mouse buttons pressed during the move event (only
supports Left, Mid, and Right mouse buttons). The double click
event appends a single argument, the mouse button that was
double clicked.

* Beautified and added copyright

* Add QEvent data to TLabel callbacks

This commit reverts previous changes in TLabel.cpp that passed
additional arguments to the appropriate event handler function.
Instead, that functionality was moved to TLuaInterpreter. An
additional argument was declared for
TLuaInterpreter::callEventHandler, allowing for the passing of
a QEvent. This argument was given a default value to maintian
compatibility.

Now, in TLuaInterpreter::callEventHandler, the function
processes the passed in event for relevant information,
pushing it on to the Lua stack as a single Lua table. This
table is appended to the current list of parameters passed back
through the event function, which prevents breaking existing
Mudlet scripts.