Revise: add (non-Mono) Ubuntu fonts, inclusion control & better licence display
Adds the missing code to store the Ubuntu fonts into a resource file; but
given that Linux distro package creators may already have the fonts in
their native package range (or they do not accept the font as one they
can distribute) steps have been taken to separate out the font resources
into a separate resource file that can be omitted at build time with
an environment variable "WITH_FONTS" set to have the value "NO".

As it happens, at the time of writing the Ubuntu fonts had just been
lined up to be placed into the "non-free" area of the Debian "unstable"
version (it has been waiting around seven years because the Ubuntu Font
Licence was just restrictive enough not to be acceptable to the DFSG -
it required a certain form of naming of *modified* versions and that seemed
to impinge on a freedom that Debian demands for "Free" software). Ideally
I would have liked to not have the font files embedded directly into the
source code and instead arranged for them to be obtained from upstream
(currently:
https://launchpad.net/ubuntu/+archive/primary/+files/
                              ubuntu-font-family-sources_X.YY.orig.tar.gz )
directly but only if they were wanted.

This commit adds ALL the remaining (non-Monospaced) Ubuntu fonts from the
original source to Mudlet.

As there is a need to include the specific licence file for these fonts in
our "About Dialog" but only if they were required - this justified my
importing code similar to that which I will be using to provide
translations of the GUI part of the application.  Specifically it moves
the HTML used for the textual content for the licence and about details
from the UI form file to a C++ source file. This has to happen to allow
for variations in the content to depend on build and/or run-time conditions
and also makes it possible avoid the horrendous obfuscation that Qt
Designer wreaks on any HTML it gets it's hand on.  By using a modular
coding style it was possible to boiler-plate multiple instances of various
free but non-GPL licences that we must include - and to omit things that
are not required for different build configurations, such as not including
the licences for the recently added updater components on builds that do
not include them.

Note that the line:
Q_INIT_RESOURCE(mudlet);
should not be included in main.cpp as that macro is only needed for Qt
library projects and not application ones - it has been removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>