BugFix: ensure Central Debug Console closes at applicantion exit
It seems that a recent change (the addition of the updater code) may have
included code that modifies the
(bool)QtGUIApplication::quitOnLastWindowClosed
flag and that and/or the removal of the qApp->quit() call at the bottom
of (void) mudlet::closeEvent(QCloseEvent*) seems to be why the Central
Debug Console - if visible - prevents the Mudlet application from shutting
down.  As far as I can tell the previous mudlet::closeEvent() only closed
the TConsole instance inside the Console but it did not close the
QMainWindow {held in (QMainWindow*) mudlet::mpDebugArea} itself.

This commit seems to fix the issue by using the close() method of the
QMainWindow rather then what is it's central widget (the TConsole that has
the isDebugConsole flag set).  For safety's sake I have converted the
mudlet::mpDebugArea pointer from a T* to a QPointer<T> form, so that a
test for it's continued existence in the mudlet::closeEvent() is more like
to work in unanticipated or unusual situations...!

Similarly I also ensure that the Qt::WA_DeleteOnClose flag is set on the
QMainWindow of the Central Debug Console just before we close it - just to
make absolutely sure it dies...! }8->

This should close #1435 "Mudlet application does not shut-down when Central
Debug Console is active".

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>