BugFix: ensure Updater option control works when no profile loaded
The code that initialises the control of the updater must be initialised
so that development versions do not have it enabled.  Unfortunate because
two recent PRs (one of mine to allow the Profile Preferences to operate
safely without a profile loaded and one of Vadim's to add the updater)
both modified chunks of the dlgProfilePrefences class the merge algorithm
allowed the new control for the updater to be placed in amongst other
code that was shifted to only work when there WAS a profile loaded when
in fact it was an exceptional piece that should have stayed in the
constructor.  As the code that had to be relocated was around half of
a chunk that seems to have been hived off into a separate method
(void) dlgProfilePreferences::loadSpecialSettingsTab()
it was cleaner to re-position the remaining code back into the single
place it was called from and remove the function completely.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>