Enhance: add updater inclusion control to qmake project file
When I tried to do this in the PR that introduced the updater code I think
the reason my qmake implimentation failed was because I tried to include
the sparkle glue code directory - which is not what is needed (the use of
the code in that directory is actually handled as part of the inclusion of
the sparkle code {MacOS only} already - and including the directory just
"broke" things.

In order to enable better platform specific qmake project files in the
future I have added some temporary test and messages investigating the
qmake "platform" "scope" values to the project file and I want to see what
the results are when run on the CI platforms.  These scope variables are
not properly detailed anywhere that I have found on-line but I have
determined through local testing that "linux"; "linux-clang" and
"linux-g++" work - which is already an addition to our knowledge on the
subject - further testing with the CI platforms have provided a few more
but I have left in place some unconfirm ones, together with a request
for anyone finding them to report back to me...!

Also in the qmake mudlet.pro file I fixed some surplus ';'s.

Modify Travis default that pulled in all git submodules automatically - as
we now do it selectively ourselves.  This included a Windows only
requirement for qmake builds to include the lua code formatter - where
we use the AppVeyor CI to also generate binaries for distribution.
I would recommend everyone else to use luarocks to get this and some
other run-time lua requirements and to this end only an AppVeyor CI
build should attempt to install the got submodule...

I found that I hade to exclude some further Windows code in non-updater
equipped builds in main:main!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>