Revise: lua code reformatter matters and add missing CMake "make install"
* Include lua code reformatter git submodule in all cases
* Add Lua Code Formatter files to qmake *nix "make install"
* Provide "make install" action for CMake builds on *nix.

It isn't obvious but we use a modified (for lua 5.1 amongst other things)
lua_code_formatter ("lcf") lua module which we hold in our own Github repo.
This is compiled on-demand within:
(void) TLuaInterpreter::initIndenterGlobals()
and we must ensure our specified branch's source code is present at
run-time.

Without this an "[ ERROR ]" message appears on profile loading because the
files are not included in the shared read-only file-set for a "proper"
*nix installation.

Note - the CMake build does not yet include the unix "make install"
capability.

Also had to rename lua_code_formatter to lcf so that the lua module can be
loaded. Closes Issue #1471 .

Renamed loading messages to "lua code formatter" to better match
"Format all" invoking context menu action in edbee widget.

I did encounter one issue in that the compiled with CMake project system,
the binary mudlet did not have an idea where the non-default Qt libraries
which it was compiled against (in a /opt/Qt/5.8/gcc_64/lib location) was
not found at run-time and it failed to run (because my system Qt
installation is not new enough to have the required Qt5Gamepad library)
unless it had a LD_LIBRARY_PATH environmental variable set to
"/opt/Qt/5.8/gcc_64/lib" before it is run. I think this is not unexpected
and does not represent an actual failing, but is something to remember in
when "install"-ing mudlet in a *nix system {it won't be am issue for
packager because it will be using and compiled against the system provided
libraries in that situation}.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>