Enhance: add updater inclusion control to qmake project file (#1457)
Enhance: add git submodule inclusion control to qmake/cmake project files...

... AND add Unix 'make install' support to latter"

Originally called:

* Enhance: add updater inclusion control to qmake and cmake project file

Combined summary of all squashed commit messages in original commits:

When I tried to do this in the PR that introduced the updater code I think
the reason my qmake implementation failed was because I tried to include
the sparkle glue code directory - which is not what is needed (the use of
the code in that directory is actually handled as part of the inclusion of
the sparkle code {MacOS only} already - and including the directory just
"broke" things.

In order to enable better platform specific qmake project files in the
future I have added some temporary test and messages investigating the
qmake "platform" "scope" values to the project file and I want to see what
the results are when run on the CI platforms.  These scope variables are
not properly detailed anywhere that I have found on-line but I have
determined through local testing that "linux"; "linux-clang" and
"linux-g++" work - which is already an addition to our knowledge on the
subject - further testing with the CI platforms have provided a few more
but I have left in place some unconfirm ones, together with a request
for anyone finding them to report back to me...!

Also in the qmake mudlet.pro file I fixed some surplus ';'s.

I've modified both the Travis and AppVeyor defaults that pulled in all
git submodules automatically - as we now do it selectively ourselves.

It isn't obvious but we use a modified (for lua 5.1 amongst other things)
lua_code_formatter ("lcf") lua module which we hold in our own Github repo.
This is compiled on-demand within:
(void) TLuaInterpreter::initIndenterGlobals()
and we must ensure our specified branch's source code is present at
run-time.

Without this an "[ ERROR ]" message appears on profile loading because the
files are not included in the shared read-only file-set for a "proper"
*nix installation.

Also had to rename lua_code_formatter to lcf so that the lua module can be
loaded. Closes Issue #1471 .

Renamed loading messages to "lua code formatter" to better match
"Format all" invoking context menu action in edbee widget.

I found that I had to exclude some further Windows code in non-updater
equipped builds in main:main!

This also provides CMake "make install" support for unix builds so that
it mirrors the QMake provision - if done within QtCreator it does take a
bit of fiddling around to use "sudo" instead of "make" with arguments of
"-A sh -c 'make install'" and something specified in the sudo configuration
(or set with a SUDO_ASKPASS) to get the password needed via a GUI
widget.

Added Lua Code Formatter files to qmake AND cmake *nix "make install".

Also reworded some comments in the project files in the light of peer and
self review.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>