BugFix: correct fore/back ground being swapped in setTextFormat (Option A)
=========== ALTERNATIVE A: Correct code to match documentation ===========

The code reported errors in the arguments to the Lua function as per the
Wiki documentation at the time - but the arguments were used the wrong
way around so that the background colour specified was used for the
foreground.

This will break things for people who are already using the code and have
already found out that the documentation was wrong.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>