BugFix: correct fore/back ground being swapped in setTextFormat (Option B)
=========== ALTERNATIVE B: Correct error messages to match code ===========

The code reported errors in the arguments to the Lua function as per the
Wiki documentation at the time - but both were wrong in regard to how the
arguments were actually used.

This will not break things for people who have already found out that the
documentation was wrong and have ignored the discrepancies but it does
fix the error messages to reflect how the code actually works.

The Wiki will need to be revised - possibly with an explanation of defect
in older version help texts.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>