Enhance: add update available indication to main toolbar (Linux/Windows)
This is my attempt to address issue #1483 .

This modifies the "About Mudlet" toolbar button to make it a menu (and
changes the icon) should updates be set NOT to be downloaded automatically
and there is at least one available.  The menu action has the original
"About Mudlet" action and a "Review update..." action which allows the user
to go to the same place as if they had clicked on the "Check for updates"
main menu item.

By necessity this commit expands the CI build matrix to check for
compilation without the updater code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>