BugFix: remove unneeded connect/disconnect(...) for Search Engine selector
I think this was a merge error, but these are no longer needed as the
selection of search engine is now determined when the profile preferences
dialog is closed instead of being set immediately the control is adjusted.

The effects of them being present are run-time error messages because the
specified slot function is no longer needed or present in the code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>