Enhance: tweaks needed to get FreeBSD build working
I have only managed to get it to build with the default clang 4.0.0 on my
FreeBSD 11.1-RELEASE amd64 platform. (The various gcc compiler versions do
compile but fail during initialisation of the CRT).  Conversely I could not
get the lldb 4.0.0 debugger to work with Qt Creator 4.5.0 (and Qt 5.7.1)
but the "ports" gdb801 (8.0.1) works find provide there is a "python"
around.  Note that the default gdb is (6.1.1) is too old to work with Qt
Creator as it does not have the needed python bindings - apparently.

As hunspell is likely built from "ports" (as will Mudlet, eventually 8-) )
we need to look for it in the /usr/local/... area of the FSH rather than
the /usr/... part that "pkg" managed "system" components will use.

I have also followed the recommendation to use UTF-8 encoding for the,
admitted short pathFileName specification for the paths for Windows...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>