Revise: make package exporter information text reflect previous options
When entering the name for the new package/module there are two check-box
options:

1) "Retain staged files?" (Checked by default) - controls whether the staging
area (for the files and folders to be included in the package/module
as well as the config.lua" and the "<packageName>.xml" files is:
EITHER a sub-directory of "<profileHomeDir>/tmp/" (checked option) that is
retained after the package exporter dialog is closed {as previous Mudlet
versions} did;
OR a "proper" temporary directory in the OS specific place for such things
(unchecked non-default option) that is UNCONDITIONALLY DELETED after the
package exporter dialog is closed. This new feature prevents the Mudlet
area of the file-system from accumulating the "left-overs" from previous
packet creation sessions.

2) "Overwrite existing package/module" (Unchecked by default) this control
allow the user to prevent (unchecked)/allow (checked) the finally produced
archive file that is the end product of the process to clobber an existing
file of the same name in the destination directory specified.  Whilst the
default option of NOT allowing overwriting will mess with the "sync"
process it DOES reflect previous Mudlet behaviour - perhaps that was why
it didn't use to work for package/modules that were in the "archive" .zip
or .mpackage file types.  As it happened this failing only became clear
once I started to generate error messages for this process!!!

Anyhow, the choice made in the first part of the process for the first
of these options is now used to modify the "help/information" text that is
displayed later on...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>