Enhance: add "development" splash screen for non-Release versions
This would be an "Easter egg" except that the way to make it visible is not
something that the user can do but instead it is used whenever the code
detects that it is running on a version where the APP_BUILD prefix is the
default "-dev". Originally it was going to be if that prefix is NOT empty
but that would mess things up for packagers who, quite rightly, mark
binaries they modify for their own purposes with something to destinguish
it from ones that have been downloaded from our website.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>