BugFix: get command line arguments for help and version working on Windows
This revision runs a GUI application on WINDOWS ONLY when the -h/--help or
-v/--version command line options are given so that a QMessageBox can be
used to display the information shown on the command line for other
platforms.  This is because Windows makes it very hard/impossible to have
an application that can be both run on the command line and as a GUI
application.

This also reverts the "command line argument introducer" back to be '-' or
'--' (the same as the other platforms) because that is what Qt applications
use, even on Windows, where this was normally '/'...


Signed-off-by: Stephen Lyons <slysven@virginmedia.com>