Revise: attempt to improve formatting of Windows command line "--help" text
In order to get the best quality layout for this I had to change to using
an HTML/RichText table - and at that point I realised that future
(G)UI translation efforts would also have to handle this (I have a workable
scheme in mind to do this in the future, despite this text being required
without the main application having started up and installed the needed
translations) so I broke the required HTML up in such a way that only the
textual content will need to be translated without the HTML tags!

As I was not developing this on the OS platform where it will be used I
added a C pre-processor symbol "DEBUG_TEST_WIN_CMDLINE_TEXTS" that can be
defined to cause the other OS to behave as if it was Windows JUST to
display these dialogues.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>