Revise: get LuaGlobal.lua to handle Windows path names (#1509)
Was bugging out for me as the testing for "prefixes" in LuaGlobal.lua was
broken for me when running from a directory a level down in a shadow build
on a Windows platform. The test for "lua.attributes(prefixes[i])"
effectively meant does this exist in the file system, and the only one that
did pass that test for me was the plain "/" one.  I have revised the
process so that on the Windows platform all paths are changed to use '\\'
and on all others '/' is used. Also the test now looks to find the
LuaGlobal.lua file (the one that is actually being executed at the time) so
that it can determine the path to use to load in the other Mudlet & Geyser
files.  By setting the debugLoading variable to true various debugging
messages will be displayed.

WARNING: This commit was prepared on my Windows PC and I am not sure it has
the correct EOL handling.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>