Revise: fixup incorrect Q_OS_XXXX values {Main part}
Q_OS_WIN includes Q_WINRT which is not what we want - use Q_OS_WIN32
(which is also set on 64-bit windows).

Q_OS_MAC is a deprecated synonym for Q_OS_DARWIN, and Qt documentation says
"Do not use."

There are three places that are NOT fixed by this commit however they will
clash with a pending PR #1506 that needs to modify the same lines so will
be fixed by a commit in that set of changes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>