Refactor: reorganise Replay execution flow
This is partly to fix the previous commits that did not accommodate replays
initiated via the loadRawFile lua command. The program flow has been
streamlined to avoid an unneeded excursion into the TConsole class, instead
all replays are started via a
(bool) mudlet::loadReplay(Host*, const QString&, QString*)
method that deduces whether the replay was started from the lua system by
a non-null third argument and NOW handles both a absolute and an a relative
path and file name (relative to the ./logs subdirectory of the inititing
Host).

Also note:
* renamed (bool) mudlet::isControlless() to
  (bool) mudlet::isControlsVisible() - {with inverted logic} so the name is
  "better" according to peer review.
* renamed (void) ctelnet::_loadReplay() to
  (void) ctelnet::loadReplayChunk() - which better explains its purpose and
  avoids a method beginning with a '_' which seems a bit iffy to me in
  C/C++!
* renamed (void) ctelnet::readPipe() to
  (void) ctelnet::slot_processReplayChunk() - which better explains its
  purpose and notes its SLOTness (it is wired up to an implicit QTimer).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>