Enhance: add ability to set any glyph as a room symbol
(Actually it is possible to use a short word as well as anything printable
from any of the Unicode Multiple Planes.)

Adds controls to the profile preference to set the (preferred) font to
use to set the room symbols from and a checkbox to only use that font.
Additionally a sub-dialog can be brought up which lists the details of all
the different symbols on the map - showing the Unicode codepoint(s) for
each and showing how they would be rendered if only the selected font is
used and if any font is permitted, along with a count of the usages and
the rooms that use each one...  A status icon is also displayed showing
whether the symbol can be rendered entirely with the selected font (green
tick), only by using glyphs from other fonts (yellow ! warning) or not with
the current fonts on the system (red/white cross).  This allows a user to
make a sensible selection of a font to use or whether they will have a
problem (and a replacement by the replacement character '�') for any
symbols.

Updates Lua setRoomChar and getRoomChar and adds replacement setRoomSymbol
and getRoomSymbol which handle the wider range of things that can be used.
getRoomChar will now return '?' and an error message if the symbol cannot
be represented by a single ASCII/ISO 8859-1 character - but will still work
for old maps as before - although it NOW allows for an existing character
to be cleared with an empty string or a space as the char attribute.

The map format version has been incremented to allow the data needed to be
saved directly into the binary file format but failback code is in place
that means that this feature can be carried in map and room user data
instead for map format versions down to 17 - the current default is 18 and
there is limited support to fail gracefully down to the 16 that Mudlet 2.1
uses (all the room letter markings that are not supported will become '?',
and the font data will be lost, but the correct room character data will
still be in the room user data.)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>