Revise: some tweaks suggested/needed
Replaced some colour specifications (white and transparent) with Qt
constants.

Use the same inline function flushSymbolPixmapCache() to clear the map
symbol pixmap cache in all places where it might be useful.

Simplify a couple of places where an if(...) {...} else {...} can be
replaced with the (...) ? (...) : (...) operator.

Limit the number of room numbers displayed for each symbol in the new
widget - to avoid complications where there are huge numbers of rooms using
a symbol.

Replace a use of QTableWidget::clearContents() with
QTableWidgets::setRowCount(0) as I was getting some odd, deep in the Qt
internal library issues {Fatal Seg. Faults!} with the former, which I
suspect, but could not prove, might have been a re-entrancy issue caused by
the method containing it being called indirectly by an asynchronous SIGNAL/
SLOT originating in the value change from the map symbol font selection
QFontComboBox...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>