Revise: improve non-XY exits/stub exits/doors indications on 2D Map
I found that the existing triangular markings drawn on the room symbols to
indicate non-XY exits (i.e. up, down, in and out) mask any glyph drawn on
as a symbol because for the real exits they are completely filled in black.
This commit enhances the design so that a mid shade "brush" is used instead
for real exits and a light cross-hatch brush is used for stub exits.
Importantly the colour used for the brush is either a contrasting black or
white - the same as the pen used to draw the outline but it is now drawn
in the same colour as the door markings on XY plain exits - so that doors
can now be indicated on those up/down/in/out exits or stubs.

The tool-tips in the room exits dialogue have been amended to suit this
long-planned (by me) enhancement!

The pen used to draw the outline is also made a non-zero width size.

Stub in and out exits were not being properly drawn in the prior code as
there was a code limitation that ended up just drawing a horizontal line.
Now both stub and real exits of these types are now drawn as a PAIR of
triangles, with IN exits being an inward pointing pair each around half
the size of the prior one, and OUT exits being an outward pointing pair
outside the IN one.  I think it is more intuitive what the new versions are
indicating...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>