Enhance: tweaks needed to get FreeBSD build working (#1506)
This commit is squashed down from several from the pull-request,
in summary:

I have only managed to get it to build with the default clang 4.0.0 on my
FreeBSD 11.1-RELEASE amd64 platform. (The various gcc compiler versions do
compile but fail during initialisation of the CRT).  Conversely I could not
get the lldb 4.0.0 debugger to work with Qt Creator 4.5.0 (and Qt 5.7.1)
but the "ports" gdb801 (8.0.1) works find provide there is a "python"
around.  Note that the default gdb is (6.1.1) is too old to work with Qt
Creator as it does not have the needed python bindings - apparently.

As hunspell is likely built from "ports" (as will Mudlet, eventually 8-) )
we need to look for it in the /usr/local/... area of the FSH rather than
the /usr/... part that "pkg" managed "system" components will use.

I have also followed the recommendation to use UTF-8 encoding for the,
admitted short pathFileName specification for the paths for Windows...

Revised to allow command line argument processing to work on.
FreeBSD, this will enable the -h/--help, -v/--version and -q/--quiet
command line arguments to also work for FreeBSD.

Revised to add FreeBSD note to GitHub front page readme file.

Also add some non-build files to qmake project file so they show up in
Qt Creator ID

Revised to fixup incorrect Q_OS_XXXX values {Enhance_tweaksForFreeBSD part}:

Q_OS_WIN includes Q_OS_WINRT which is not what we want - use Q_OS_WIN32
(which is also set on 64-bit windows).

Q_OS_MAC is a deprecated synonym for Q_OS_DARWIN, and Qt
documentation says "Do not use."

This touches the three places that are mentioned as being omitted in PR #1532 .

Revised to add dis-owner for FreeBSD from all but myself -  this was
requested to make it clear that if Mudlet works for FreeBSD then
that is great, but if it breaks, then developers other than myself are not
necessarily going to be able or willing to put the pieces back together.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>