BugFix: prevent multiple replays from running + main menu/toolbar loss (#1519)
This commit is a squashed down one from the original PR, in summary:

The replay system, because it attaches a temporary toolbar to the main
menu needs to be restricted to only have one replay active at a time
- otherwise things *break*...

This commit overhauls the replay system to do that - and also properly
clean up the Widgets that are created for each replay but were then not
previously being destroyed.  To reinforce/enable the one replay limit the
menu and toolbar buttons are "locked" down and disabled when a replay is in
use.  A corner case where multiple profiles are loaded and a profile with
a replay running is closed before the replay ends is correctly handled to
reset and unlock the replay system for the remaining profile(s).

During debugging of enabling and disabling of the replay controls I found
that the existing system to prevent loss of access to one of either the
main toolbar or the application menu-bar was not respecting the user
settings when a second or subsequent profile was loaded. I have "improved"
the system now to allow one but not BOTH of the main menu and the main
toolbar to be hidden after a profile is loaded.

Once a profile has been loaded and both the main menu and the main toolbar
has been hidden if those options have been chosen then a pair of options
will appear on all TConsole context menus to return either of them to be
shown "Always" - although there is also a Qt system provided clickable
option on the context menu for any "UserWindow" or floating toolbar
that can also show/hide the (currently anonymous) main toolbar...!

The above revealed yet another "issue" in that the context menu had
StatusBar texts set to be displayed but we no longer use/have a status-bar
on the main window - so I have converted the ones in that context menu to
be ToolTips instead.

Refactored to reorganise Replay execution flow

This is partly to fix the previous commits that did not accommodate replays
initiated via the loadRawFile lua command. The program flow has been
streamlined to avoid an unneeded excursion into the TConsole class, instead
all replays are started via a
(bool) mudlet::loadReplay(Host*, const QString&, QString*)
method that deduces whether the replay was started from the lua system by
a non-null third argument and NOW handles both a absolute and an a relative
path and file name (relative to the ./logs subdirectory of the inititing
Host).

Also note:
* renamed (bool) mudlet::isControlless() to
  (bool) mudlet::isControlsVisible() - {with inverted logic} so the name is
  "better" according to peer review.
* renamed (void) ctelnet::_loadReplay() to
  (void) ctelnet::loadReplayChunk() - which better explains its purpose and
  avoids a method beginning with a '_' which seems a bit iffy to me in
  C/C++!
* renamed (void) ctelnet::readPipe() to
  (void) ctelnet::slot_processReplayChunk() - which better explains its
  purpose and notes its SLOTness (it is wired up to an implicit QTimer).

Fixed to label and prevent main toolbar from being hidden if unsafe to do so.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>