Enhance: add GBK/GB18030 decoder for Far Eastern locales
This commit adds in the ability to process the 1/2 (or 4 for GB1830) byte
sequences that at least Mainland China uses for simplified Chinese for even
historic language symbols. It actually uses the Qt decoder but like the
UTF-8 processor in the TBuffer class it needs to process the incoming
stream byte by byte so as to map each QChar in the recovered data to a
corresponding TChar attribute element that Mudlet needs to have to work
with the visual presentation of that text data. Note that once processed
in TBuffer the textual data is held in the UTF-16BE form that Qt uses for
its QString datatype.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>